+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 
 - UI: Add country selection step after document selection
 - Internal: The release script and the `release/RELEASE_GUIDELINE.md` file now include the information needed to update the `MIGRATION.md` file.
+- UI: Added Residence Permit option for document selection
 
 ### Changed